move read and read/write of tiecdr to proper source file
